requirements.txt
Django>=5.2,<5.3
wagtail>=7.0,<7.1
python-decouple
wagtail-grapple
django-extensions
PyMySQL
python-dotenv
